+++ CHANGES.rst
@@ -1,6 +1,9 @@
 Changes in 1.0.2 (2020-07-28)
 ===============================================
- 
+
+Improvements:
+ * Implement PIN protection (#3436).
+
 Bug fix:
  * Registration: Do not display the skip button if email is mandatory (#3417).
  * NotificationService: Do not cache showDecryptedContentInNotifications setting (#3444).